requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cors==0.9.11
Django==2.2.12
django-cors-headers==3.10.0
django-filter==21.1
django-rest-swagger==2.1.2
django-utils-six==2.0
djangorestframework==3.12.2
djangorestframework-filters==0.11.1
drf-yasg==1.20.0
filelock==3.0.12
future==0.18.2
gevent==21.1.2
greenlet==1.0.0
idna==2.10
importlib-metadata==4.8.1
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
kombu==5.1.0
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==21.0
prompt-toolkit==3.0.21
psycopg2-binary==2.8.3
pyparsing==2.4.7
python-decouple==3.5
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-file==1.5.1
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
simplejson==3.17.2
six==1.15.0
soupsieve==2.2.1
sqlparse==0.3.1
tldextract==3.1.0
typing-extensions==3.10.0.2
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.10
vine==5.0.0
wcwidth==0.2.5
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.3.0
